feat: add functions

+++ src/main/kotlin/com/example/springbackend/school/domain/SchoolDomain.kt
@@ -1,10 +1,8 @@
 package com.example.springbackend.school.domain
 
-import java.math.BigDecimal
-
 class SchoolDomain(
     val schoolName: String,
-    val latitude: BigDecimal,
-    val longitude: BigDecimal,
+    val latitude: Double,
+    val longitude: Double,
     val id: String = ""
 )
